requirements.txt
customtkinter==5.2.2
openpyxl==3.1.5
pyinstaller==6.16.0
pillow==11.3.0
